fix(config): Locate configuration menu from 'MoonPhase.js' to 'App.js'
