+++ react/src/App.css
@@ -14,6 +14,32 @@ code {
     monospace;
 }
 
+.menu {
+position: absolute;
+top: 0;
+left: -400px;
+"width: 250px;"
+height: 100%;
+        background-color: #333;
+        overflow-y: auto;
+        z-index: 10;
+padding: 1rem;
+         box-sizing: border-box;
+transition: left 0.3s ease-in-out;
+display: flex;
+         flex-direction: column;
+         align-items: center;
+         padding-top: 20px;
+         padding-right: 30px;
+         z-index: 5;
+color: white;
+       font-weight: bold;
+}
+
+.menu.open {
+left: 0px;
+}
+
 .App {
   text-align: center;
 }
