+++ react/src/MoonPhase.css
@@ -97,33 +97,6 @@ cursor: pointer;
 color: white;
 }
 
-.menu {
-position: absolute;
-top: 0;
-left: -250px;
-"width: 250px;"
-height: 100%;
-        background-color: #333;
-        overflow-y: auto;
-        z-index: 10;
-padding: 1rem;
-         box-sizing: border-box;
-transition: left 0.7s ease-in-out;
-display: flex;
-         flex-direction: column;
-         align-items: center;
-         padding-top: 20px;
-         padding-right: 30px;
-         z-index: 5;
-color: white;
-       font-weight: bold;
-}
-
-.menu.open {
-left: 0;
-}
-
-
 
 /* Smaller font size for screens with a width of 768px or less */
 @media (max-width: 768px) {
