+++ react/src/timeZoneNames.js
@@ -1,8 +1,10 @@
+import moment from 'moment-timezone';
+export const timeZoneNames = moment.tz.names();
 
-export const timeZoneNames = [
-  'UTC',
-  'America/New_York',
-  'Europe/London',
-  'Asia/Tokyo',
-  // Add more timezones here
-];
+//export const timeZoneNames = [
+//  'UTC',
+//  'America/New_York',
+//  'Europe/London',
+//  'Asia/Tokyo',
+//  // Add more timezones here
+//];
